Move Pages outside Components
