+++ src/components/App/App.tsx
@@ -2,16 +2,16 @@ import styles from './App.module.css';
 import { Route, Routes, useLocation, useNavigate } from 'react-router-dom';
 import { AppHeader } from '../AppHeader/AppHeader';
 import { MainSection } from '../MainSection/MainSection';
-import Login from '../pages/Login/Login';
-import Register from '../pages/Register/Register';
-import ForgotPassword from '../pages/ForgotPassword/ForgotPassword';
-import ResetPassword from '../pages/ResetPassword/ResetPassword';
+import Login from '../../pages/Login/Login';
+import Register from '../../pages/Register/Register';
+import ForgotPassword from '../../pages/ForgotPassword/ForgotPassword';
+import ResetPassword from '../../pages/ResetPassword/ResetPassword';
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
-import Profile from '../pages/Profile/Profile';
+import Profile from '../../pages/Profile/Profile';
 import { useDispatch } from 'react-redux';
 import { useEffect } from 'react';
 import { fetchIngredients } from '../../store/actions/ingredientsListActions';
-import IngredientPage from '../pages/IngredientPage/IngredientPage';
+import IngredientPage from '../../pages/IngredientPage/IngredientPage';
 import { performRefreshTokens } from '../../store/actions/accountActions';
 
 function App(): React.ReactElement {
